Improved dependencies of modules
Any import SalesforceError should not initiate the full orchestra
