+++ salesforce/tests/test_unit.py
@@ -108,11 +108,11 @@ class TestQueryCompiler(TestCase):
 
     def test_none_method_queryset(self):
         """Test that none() method in the queryset returns [], not error"""
-        request_count_0 = salesforce.backend.query.request_count
+        request_count_0 = salesforce.backend.driver.request_count
         self.assertEqual(tuple(Contact.objects.none()), ())
         self.assertEqual(tuple(Contact.objects.all().none().all()), ())
         self.assertTrue('[]' in repr(Contact.objects.none()))
-        self.assertEqual(salesforce.backend.query.request_count, request_count_0,
+        self.assertEqual(salesforce.backend.driver.request_count, request_count_0,
                          "Do database requests should be done with .none() method")
